Misc cleanup.

+++ src/main/java/org/avp/client/render/BlastDoorPlacementBoxRenderer.java
@@ -58,9 +58,9 @@ public class BlastDoorPlacementBoxRenderer
 
                         Vec3d hitVec = ClientGame.instance.minecraft().objectMouseOver.hitVec;
 
-                        double x = p.lastTickPosX + (p.posX - p.lastTickPosX) * (double) event.getPartialTicks();
-                        double y = p.lastTickPosY + (p.posY - p.lastTickPosY) * (double) event.getPartialTicks();
-                        double z = p.lastTickPosZ + (p.posZ - p.lastTickPosZ) * (double) event.getPartialTicks();
+                        double x = p.lastTickPosX + (p.posX - p.lastTickPosX) * event.getPartialTicks();
+                        double y = p.lastTickPosY + (p.posY - p.lastTickPosY) * event.getPartialTicks();
+                        double z = p.lastTickPosZ + (p.posZ - p.lastTickPosZ) * event.getPartialTicks();
 
                         double locationX = Math.floor(hitVec.x);
                         double locationY = Math.floor(hitVec.y);
